pyproject.toml
[tool.poetry]
name = "drf-plus-react-equals-love-backend-jkube38"
version = "0.1.0"
description = ""
authors = ["Jordan Kubista <jkube29@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
